requirements.txt
tqdm
fire
numpy
scipy
datasets
pytz
matplotlib
transformers
google
google-api-python-client
google-generativeai
fairscale
sentencepiece
tiktoken
python-dotenv
huggingface_hub